solve merge issues

+++ Assets/Interfaces/REST/Resources/GeolocationResource.cs
@@ -1,3 +1,3 @@
 ﻿namespace Hampcoders.Electrolink.API.Assets.Interfaces.REST.Resources;
 
-public record GeolocationResource(decimal Latitude, decimal Longitude, int? Accuracy, string Source);
+public record GeolocationResource(double Latitude, double Longitude, int? Accuracy, string Source);
